improvmenet.

+++ Makefile
@@ -1,5 +1,5 @@
 # just a test
 install_configs:
-	for each in $$( ls -a1 | egrep -v '(Makefile|\.git|^\.\.?$$|\~$$)'); do \
-	  ln -nfs `pwd`/$$each $$HOME/$$each; \
+	for each in $$( ls -a1 | egrep -v '(Makefile|\.git|^\.\.?$$|\~$$|#)'); do \
+	  [ -e $$HOME/$$each ] || ln -nfs `pwd`/$$each $$HOME/$$each; \
 	done;
